refactor(niri): lift places and trackability out of the interpreter
Neither touches interpreter state. `place` answers what a borrow or lvalue
chain names; `trackability` answers which locals a walk may hold a value for,
as pure functions of a body and the maps. The interpreter consumes both.

Worth their own modules because the rule they encode is the one this engine is
easiest to get wrong on — a read counts wherever it appears, a write only where
the walk performs it — and both of this cycle's stale-container bugs lived
there.

Two doc blocks had come loose from what they described: the CTFE-eligibility
list that had grown onto `EditSink`, still claiming a `stores` gate that no
longer exists, and `ExecutedWrites`'s own, left above `CallRun` when that type
was inserted between them. `value_with_path` moves to `const_eval` as
`Value::with_path`, where the state-free operations on `Value` already live.
