+++ wado-compiler/src/niri.rs
@@ -349,6 +349,12 @@ pub const DEFAULT_STEP_BUDGET: u32 = 10_000;
 // Field knowledge
 // ──────────────────────────────────────────────────────────────────────────────
 
+mod place;
+mod trackability;
+
+use place::{overlapping_places, place_of};
+use trackability::{ExecutedWrites, aggregate_safe_locals, clobbered_locals};
+
 /// A dense set of local indices, backed by a bitset indexed by the local
 /// index itself.
 ///
@@ -434,22 +440,6 @@ pub struct AliasInfo {
     pub alias_groups: IndexMap<u32, IndexSet<u32>>,
 }
 
-/// Decide whether a function may be evaluated at compile time.
-///
-/// The check is a conservative pure-and-safe gate, applied once when the
-/// driving visitor builds the [`CalleeMap`]:
-///
-/// - `effects.is_empty()` — no `with` clauses (the effect system's purity
-///   witness, modulo trap effects which Wado tracks separately).
-/// - `body.is_some()` — has a Wado-source body. External / CM-import
-///   functions have no inspectable body.
-/// - `!is_cm_binding && !is_dispatch_wrapper && !is_cm_export` —
-///   synthesized ABI bridges aren't real Wado functions.
-/// - `!is_async && task_return_type.is_none()` — async functions
-///   participate in the CM async runtime; not CTFE-safe.
-/// - `stores.is_empty()` — `stores[...]` is moot for CTFE (we don't pass
-///   refs), but bail conservatively.
-///
 /// Commit sink for niri's body rewrites. The rewrite logic reads through
 /// [`EditSink::body`] and commits every edit through the sink, so two backends
 /// can share it: [`BodySink`] mutates a `Body` in place — used for throwaway
@@ -671,263 +661,11 @@ fn let_ref_global(body: &Body, stmt: &StmtKind) -> Option<(u32, GlobalKey)> {
     Some((*local_index, (module_source.clone(), name.clone())))
 }
 
-/// The local an lvalue or borrow chain roots at: `x`, `x.f`, `x[i]`, `*x`, and
-/// any nesting of those.
-fn lvalue_root_local(body: &Body, op: Operand) -> Option<u32> {
-    match &body.exprs[op.as_expr()?].kind {
-        ExprKind::Local { index, .. } => Some(*index),
-        ExprKind::FieldAccess { expr: inner, .. } | ExprKind::Index { expr: inner, .. } => {
-            lvalue_root_local(body, *inner)
-        }
-        ExprKind::Unary {
-            op: NirUnaryOp::Deref,
-            expr: inner,
-        } => lvalue_root_local(body, *inner),
-        _ => None,
-    }
-}
-
-/// The local a borrow or lvalue chain roots at, and the field path reaching
-/// into its value. [`lvalue_root_local`] answers only which local is touched;
-/// a write also needs to know where inside it lands.
-fn place_of(body: &Body, op: Operand) -> Option<(u32, Vec<u32>)> {
-    match &body.exprs[op.as_expr()?].kind {
-        ExprKind::Local { index, .. } => Some((*index, Vec::new())),
-        ExprKind::Unary {
-            op: NirUnaryOp::Ref | NirUnaryOp::MutRef | NirUnaryOp::Deref,
-            expr,
-        } => place_of(body, *expr),
-        ExprKind::FieldAccess {
-            expr, field_index, ..
-        } => {
-            let (root, mut path) = place_of(body, *expr)?;
-            path.push(*field_index);
-            Some((root, path))
-        }
-        _ => None,
-    }
-}
-
-/// How many of `places` reach storage `(root, path)` reaches. One place covers
-/// another when its path is a prefix: `c` covers `c.repr`, and `c.repr` and
-/// `c.used` cover nothing of each other.
-fn overlapping_places(places: &[(u32, Vec<u32>)], root: u32, path: &[u32]) -> usize {
-    places
-        .iter()
-        .filter(|(other_root, other_path)| {
-            *other_root == root && (other_path.starts_with(path) || path.starts_with(other_path))
-        })
-        .count()
-}
-
-/// `base` with the value at `path` replaced by `new`. `None` when the path does
-/// not reach a field the value has.
-fn value_with_path(base: &Value, path: &[u32], new: Value) -> Option<Value> {
-    let Some((head, rest)) = path.split_first() else {
-        return Some(new);
-    };
-    let updated = value_with_path(base.field(*head)?, rest, new)?;
-    base.with_field(*head, updated)
-}
-
-/// The writes a compile-time frame performs itself, so the local each one roots
-/// at keeps a current value instead of a stale one. Every other write still
-/// disqualifies its root.
-///
-/// A write counts only at statement position, where the executor runs it — one
-/// buried in a larger expression is merely reduced, so its target would keep a
-/// stale value.
 /// What running a call left behind: the value it returned, and what each `&mut`
 /// parameter holds at return, paired with the caller place it belongs in.
 struct CallRun {
     result: Lattice,
     writes: Vec<(u32, Vec<u32>, Value)>,
-}
-
-#[derive(Default)]
-struct ExecutedWrites {
-    /// Nothing can write through a read, so naming one says nothing about
-    /// whether the value stays current.
-    reads: IndexSet<ExprId>,
-    /// Exempt only inside a compile-time frame, which performs the write or
-    /// abandons the evaluation. An ordinary walk performs nothing, and a write
-    /// it steps over leaves its target stale.
-    writes: IndexSet<ExprId>,
-    place_assigns: IndexSet<ExprId>,
-}
-
-impl ExecutedWrites {
-    fn in_frame(
-        body: &Body,
-        ctfe_builtins: Option<&CtfeBuiltinMap>,
-        callees: Option<&CalleeMap>,
-    ) -> Self {
-        let mut statements: IndexSet<ExprId> = IndexSet::default();
-        for (_, stmt) in &body.stmts {
-            let op = match &stmt.kind {
-                StmtKind::Expr(op) | StmtKind::Let { value: op, .. } => *op,
-                _ => continue,
-            };
-            if let Some(e) = op.as_expr() {
-                statements.insert(e);
-            }
-        }
-        let mut writes = Self::default();
-        for e in &statements {
-            if let ExprKind::Assign { target, .. } = &body.exprs[*e].kind
-                && place_of(body, (*target).into()).is_some()
-            {
-                writes.place_assigns.insert(*e);
-            }
-        }
-        writes.collect_builtin_borrows(body, ctfe_builtins, &statements);
-        writes.collect_call_borrows(body, callees, &statements);
-        writes
-    }
-
-    /// An ordinary walk performs nothing, so only the reads survive.
-    fn outside_frame(
-        body: &Body,
-        ctfe_builtins: Option<&CtfeBuiltinMap>,
-        callees: Option<&CalleeMap>,
-    ) -> Self {
-        let reads = Self::in_frame(body, ctfe_builtins, callees).reads;
-        Self {
-            reads,
-            writes: IndexSet::default(),
-            place_assigns: IndexSet::default(),
-        }
-    }
-
-    /// The engine models a builtin call exactly, so the destination it writes
-    /// and the source it reads are equally current — where the frame runs it.
-    fn collect_builtin_borrows(
-        &mut self,
-        body: &Body,
-        ctfe_builtins: Option<&CtfeBuiltinMap>,
-        statements: &IndexSet<ExprId>,
-    ) {
-        let Some(map) = ctfe_builtins else {
-            return;
-        };
-        for (e, node) in &body.exprs {
-            let ExprKind::Call { func_id, args, .. } = &node.kind else {
-                continue;
-            };
-            let Some(builtin) = map.get(func_id) else {
-                continue;
-            };
-            if !builtin.is_write() {
-                for arg in args {
-                    self.record(body, arg.expr, Reach::Read);
-                }
-                continue;
-            }
-            if !statements.contains(&e) {
-                continue;
-            }
-            for (i, arg) in args.iter().enumerate() {
-                let reach = if i == 0 { Reach::Write } else { Reach::Read };
-                self.record(body, arg.expr, reach);
-            }
-        }
-    }
-
-    /// A shared receiver or by-value argument counts wherever it appears —
-    /// that is what carries a container through the `&self` reads `push` makes
-    /// of its own capacity. A `&mut` one counts only at statement position,
-    /// where a frame performs the write; elsewhere the projection merely reads
-    /// the call. A callee the map does not hold exempts nothing.
-    fn collect_call_borrows(
-        &mut self,
-        body: &Body,
-        callees: Option<&CalleeMap>,
-        statements: &IndexSet<ExprId>,
-    ) {
-        let Some(callees) = callees else {
-            return;
-        };
-        for (e, node) in &body.exprs {
-            let (func_id, receiver, args) = match &node.kind {
-                ExprKind::Call { func_id, args, .. } => (func_id, None, args),
-                ExprKind::MethodCall {
-                    func_id,
-                    receiver,
-                    args,
-                    ..
-                } => (func_id, Some(*receiver), args),
-                _ => continue,
-            };
-            let Some(callee) = callees.get(func_id) else {
-                continue;
-            };
-            let first_arg = usize::from(receiver.is_some());
-            if first_arg + args.len() != callee.arity() {
-                continue;
-            }
-            let at_statement = statements.contains(&e);
-            if let Some(receiver) = receiver {
-                match (callee.writes_receiver(), at_statement) {
-                    (false, _) if callee.reads_only(0) => {
-                        self.record(body, receiver, Reach::Read);
-                    }
-                    (true, true) => self.record(body, receiver, Reach::Write),
-                    (false, _) | (true, false) => {}
-                }
-            }
-            for (i, arg) in args.iter().enumerate() {
-                let index = first_arg + i;
-                match (callee.writes_param(index), at_statement) {
-                    (false, _) if callee.reads_only(index) => {
-                        self.record(body, arg.expr, Reach::Read);
-                    }
-                    (true, true) => self.record(body, arg.expr, Reach::Write),
-                    (false, _) | (true, false) => {}
-                }
-            }
-        }
-    }
-
-    fn reaches(&self, body: &Body, op: Operand) -> bool {
-        let mut op = op;
-        loop {
-            let Some(e) = op.as_expr() else {
-                return false;
-            };
-            match &body.exprs[e].kind {
-                ExprKind::Unary {
-                    op: NirUnaryOp::Ref | NirUnaryOp::MutRef | NirUnaryOp::Deref,
-                    expr,
-                } => op = *expr,
-                _ => return self.reads.contains(&e) || self.writes.contains(&e),
-            }
-        }
-    }
-
-    /// Records the place `op` names, peeling the borrow it may be wrapped in.
-    fn record(&mut self, body: &Body, op: Operand, reach: Reach) {
-        let Some(e) = op.as_expr() else {
-            return;
-        };
-        match &body.exprs[e].kind {
-            ExprKind::Unary {
-                op: NirUnaryOp::Ref | NirUnaryOp::MutRef | NirUnaryOp::Deref,
-                expr,
-            } => self.record(body, *expr, reach),
-            _ => {
-                match reach {
-                    Reach::Read => self.reads.insert(e),
-                    Reach::Write => self.writes.insert(e),
-                };
-            }
-        }
-    }
-}
-
-#[derive(Clone, Copy)]
-enum Reach {
-    Read,
-    Write,
 }
 
 /// Every expression id reachable from the body root, in arena order.
@@ -948,117 +686,6 @@ pub(crate) fn reachable_exprs(body: &Body) -> Vec<ExprId> {
     let mut collect = Collect(Vec::new());
     collect.visit_node(body, NodeRef::Block(body.root));
     collect.0
-}
-
-/// Locals of `body` that may bind an aggregate constant: every mention only
-/// reads the value, and no store target, mutable borrow, method receiver, or
-/// mutable argument roots at them. A write in `writes` is exempt: the executor
-/// performs it, so the value it leaves is current.
-///
-/// The read positions are listed rather than inferred from the absence of the
-/// others, so a node kind nobody taught this walk about costs a fold and never
-/// a wrong one. Under value semantics that list is long: returning, binding,
-/// storing and composing all copy.
-///
-/// Only the reachable body is scanned: a mention orphaned by an earlier rewrite
-/// cannot run, so it must not disqualify a local.
-fn aggregate_safe_locals(body: &Body, writes: &ExecutedWrites) -> LocalSet {
-    fn disqualify_root(body: &Body, op: Operand, set: &mut LocalSet) {
-        if let Some(index) = lvalue_root_local(body, op) {
-            set.insert(index);
-        }
-    }
-    fn read_value(op: Operand, reads: &mut IndexSet<ExprId>) {
-        if let Some(e) = op.as_expr() {
-            reads.insert(e);
-        }
-    }
-    let mut value_reads: IndexSet<ExprId> = IndexSet::default();
-    let mut local_mentions: Vec<(ExprId, u32)> = Vec::new();
-    let mut disqualified = LocalSet::default();
-    for e in reachable_exprs(body) {
-        let node = &body.exprs[e];
-        match &node.kind {
-            ExprKind::Local { index, .. } => local_mentions.push((e, *index)),
-            ExprKind::FieldAccess { expr, .. }
-            | ExprKind::Match { expr, .. }
-            | ExprKind::Switch {
-                scrutinee: expr, ..
-            } => read_value(*expr, &mut value_reads),
-            ExprKind::Index { expr, index } => {
-                read_value(*expr, &mut value_reads);
-                read_value(*index, &mut value_reads);
-            }
-            ExprKind::TupleLiteral { elements } | ExprKind::ArrayLiteral { elements } => {
-                for element in elements {
-                    read_value(*element, &mut value_reads);
-                }
-            }
-            ExprKind::StructLiteral { fields, .. } => {
-                for field in fields {
-                    read_value(field.value, &mut value_reads);
-                }
-            }
-            ExprKind::Assign { target, value } => {
-                read_value(*value, &mut value_reads);
-                if !writes.place_assigns.contains(&e) {
-                    disqualify_root(body, (*target).into(), &mut disqualified);
-                }
-            }
-            ExprKind::Unary {
-                op: NirUnaryOp::MutRef,
-                expr,
-            } => {
-                if !writes.reaches(body, *expr) {
-                    disqualify_root(body, *expr, &mut disqualified);
-                }
-            }
-            ExprKind::MethodCall { receiver, args, .. } => {
-                if !writes.reaches(body, *receiver) {
-                    disqualify_root(body, *receiver, &mut disqualified);
-                }
-                for arg in args {
-                    if !arg.is_mut {
-                        read_value(arg.expr, &mut value_reads);
-                    } else if !writes.reaches(body, arg.expr) {
-                        disqualify_root(body, arg.expr, &mut disqualified);
-                    }
-                }
-            }
-            ExprKind::Call { args, .. } => {
-                for arg in args {
-                    if !arg.is_mut {
-                        read_value(arg.expr, &mut value_reads);
-                    } else if !writes.reaches(body, arg.expr) {
-                        disqualify_root(body, arg.expr, &mut disqualified);
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-    for (_, stmt) in &body.stmts {
-        match &stmt.kind {
-            StmtKind::Return { value: Some(op) } => read_value(*op, &mut value_reads),
-            StmtKind::Let { value: op, .. } | StmtKind::Expr(op) => {
-                read_value(*op, &mut value_reads);
-            }
-            _ => {}
-        }
-    }
-    value_reads.extend(writes.reads.iter().chain(&writes.writes).copied());
-    for (e, index) in &local_mentions {
-        if !value_reads.contains(e) {
-            disqualified.insert(*index);
-        }
-    }
-    let mut safe = LocalSet::default();
-    for (_, index) in local_mentions {
-        if !disqualified.contains(index) {
-            safe.insert(index);
-        }
-    }
-    safe
 }
 
 fn local_binds_to_global_ref(body: &Body, local: u32, key: &GlobalKey) -> bool {
@@ -2644,7 +2271,7 @@ impl<'a> Interpreter<'a> {
             return None;
         };
         let target = path.iter().try_fold(&container, |v, i| v.field(*i))?;
-        let updated = value_with_path(&container, path, update(target)?)?;
+        let updated = container.with_path(path, update(target)?)?;
         self.bind_ctfe_local(root, Lattice::Const(updated));
         Some(())
     }
@@ -3210,60 +2837,6 @@ impl LoopSnapshot {
             body.blocks[*b].stmts.clone_from(stmts);
         }
     }
-}
-
-/// Locals of `body` that a compile-time frame cannot track, because something
-/// other than a write it performs itself can reach them: a borrow, a mutable
-/// argument, a method receiver, or an assignment buried inside a larger
-/// expression.
-///
-/// Only the reachable body is scanned, as in [`aggregate_safe_locals`]: a
-/// borrow orphaned by an earlier rewrite cannot run, and the arena keeps every
-/// node an in-place rewrite displaced.
-fn clobbered_locals(body: &Body, writes: &ExecutedWrites) -> LocalSet {
-    fn disqualify(body: &Body, op: Operand, set: &mut LocalSet) {
-        if let Some(index) = lvalue_root_local(body, op) {
-            set.insert(index);
-        }
-    }
-    let mut set = LocalSet::default();
-    for e in reachable_exprs(body) {
-        let node = &body.exprs[e];
-        match &node.kind {
-            ExprKind::Assign { target, .. } => {
-                if !writes.place_assigns.contains(&e) {
-                    disqualify(body, (*target).into(), &mut set);
-                }
-            }
-            ExprKind::Unary {
-                op: NirUnaryOp::Ref | NirUnaryOp::MutRef,
-                expr,
-            } => {
-                if !writes.reaches(body, *expr) {
-                    disqualify(body, *expr, &mut set);
-                }
-            }
-            ExprKind::MethodCall { receiver, args, .. } => {
-                if !writes.reaches(body, *receiver) {
-                    disqualify(body, *receiver, &mut set);
-                }
-                for arg in args.iter().filter(|a| a.is_mut) {
-                    if !writes.reaches(body, arg.expr) {
-                        disqualify(body, arg.expr, &mut set);
-                    }
-                }
-            }
-            ExprKind::Call { args, .. } => {
-                for arg in args.iter().filter(|a| a.is_mut) {
-                    if !writes.reaches(body, arg.expr) {
-                        disqualify(body, arg.expr, &mut set);
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-    set
 }
 
 /// The locals a matched pattern binds, paired with the scrutinee sub-values
